requirements.txt
numpy==1.24.2
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
darkdetect==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fake-useragent==1.1.1
Faker==17.0.0
frozenlist==1.3.3
gunicorn==20.1.0
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
ply==3.11
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.4.0
